+++ src/shared/Utilities/Runtime.lua
@@ -1,0 +1,9 @@
+--!strict
+local task = task
+if not task then
+	task = require("@lune/task") :: any
+end
+
+return {
+	task = task,
+}